removed RobustGPHyperconfig, it now uses SingleTaskGPHyperconfig
