+++ bofire/data_models/surrogates/single_task_gp.py
@@ -26,6 +26,7 @@ from bofire.data_models.priors.api import (
     THREESIX_NOISE_PRIOR,
     THREESIX_SCALE_PRIOR,
     AnyPrior,
+    AnyPriorConstraint,
 )
 from bofire.data_models.surrogates.trainable import Hyperconfig
 from bofire.data_models.surrogates.trainable_botorch import TrainableBotorchSurrogate
@@ -44,6 +45,8 @@ class SingleTaskGPHyperconfig(Hyperconfig):
             CategoricalInput(key="ard", categories=["True", "False"]),
         ],
     )
+    lengthscale_constraint: Optional[AnyPriorConstraint] = None
+    outputscale_constraint: Optional[AnyPriorConstraint] = None
     target_metric: RegressionMetricsEnum = RegressionMetricsEnum.MAE
     hyperstrategy: Literal[
         "FractionalFactorialStrategy", "SoboStrategy", "RandomStrategy"
@@ -53,12 +56,32 @@ class SingleTaskGPHyperconfig(Hyperconfig):
     def _update_hyperparameters(
         surrogate_data: "SingleTaskGPSurrogate",
         hyperparameters: pd.Series,
+        outputscale_constraint: Optional[AnyPriorConstraint] = None,
+        lengthscale_constraint: Optional[AnyPriorConstraint] = None,
     ):
-        def matern_25(ard: bool, lengthscale_prior: AnyPrior) -> MaternKernel:
-            return MaternKernel(nu=2.5, lengthscale_prior=lengthscale_prior, ard=ard)
+        def matern_25(
+            ard: bool,
+            lengthscale_prior: AnyPrior,
+            lengthscale_constraint: AnyPriorConstraint,
+        ) -> MaternKernel:
+            return MaternKernel(
+                nu=2.5,
+                lengthscale_prior=lengthscale_prior,
+                lengthscale_constraint=lengthscale_constraint,
+                ard=ard,
+            )
 
-        def matern_15(ard: bool, lengthscale_prior: AnyPrior) -> MaternKernel:
-            return MaternKernel(nu=1.5, lengthscale_prior=lengthscale_prior, ard=ard)
+        def matern_15(
+            ard: bool,
+            lengthscale_prior: AnyPrior,
+            lengthscale_constraint: AnyPriorConstraint,
+        ) -> MaternKernel:
+            return MaternKernel(
+                nu=1.5,
+                lengthscale_prior=lengthscale_prior,
+                lengthscale_constraint=lengthscale_constraint,
+                ard=ard,
+            )
 
         if hyperparameters.prior == "mbo":
             noise_prior, lengthscale_prior, outputscale_prior = (
@@ -82,22 +105,30 @@ class SingleTaskGPHyperconfig(Hyperconfig):
 
         if hyperparameters.kernel == "rbf":
             base_kernel = RBFKernel(
-                ard=hyperparameters.ard, lengthscale_prior=lengthscale_prior
+                ard=hyperparameters.ard,
+                lengthscale_prior=lengthscale_prior,
+                lengthscale_constraint=lengthscale_constraint,
             )
         elif hyperparameters.kernel == "matern_2.5":
             base_kernel = matern_25(
-                ard=hyperparameters.ard, lengthscale_prior=lengthscale_prior
+                ard=hyperparameters.ard,
+                lengthscale_prior=lengthscale_prior,
+                lengthscale_constraint=lengthscale_constraint,
             )
         elif hyperparameters.kernel == "matern_1.5":
             base_kernel = matern_15(
-                ard=hyperparameters.ard, lengthscale_prior=lengthscale_prior
+                ard=hyperparameters.ard,
+                lengthscale_prior=lengthscale_prior,
+                lengthscale_constraint=lengthscale_constraint,
             )
         else:
             raise ValueError(f"Kernel {hyperparameters.kernel} not known.")
 
         if hyperparameters.scalekernel == "True":
             surrogate_data.kernel = ScaleKernel(
-                base_kernel=base_kernel, outputscale_prior=outputscale_prior
+                base_kernel=base_kernel,
+                outputscale_prior=outputscale_prior,
+                outputscale_constraint=outputscale_constraint,
             )
         else:
             surrogate_data.kernel = base_kernel
